+++ tests/unit/tad/Codeception/SnapshotAssertions/BaseVisitorTest.php
@@ -1,0 +1,50 @@
+<?php namespace tad\Codeception\SnapshotAssertions;
+
+use PHPUnit\Framework\AssertionFailedError;
+
+class BaseVisitorTest extends BaseTestCase
+{
+    /**
+     * It should allow adding a visitor
+     *
+     * @test
+     */
+    public function should_allow_adding_a_visitor()
+    {
+        $removeHashEntry = static function ($jsonString) {
+            return json_encode(array_diff_key(json_decode($jsonString,true),array_flip(['hash'])));
+        };
+        $dataVisitor = static function ($expected, $current) use ($removeHashEntry) {
+            return array_map($removeHashEntry, [$expected, $current]);
+        };
+
+        $hash = md5(microtime());
+        $newHash = md5(microtime());
+
+        $this->assertNotEquals($hash, $newHash);
+
+        $jsonOne = json_encode(['test_one' => 'one', 'hash' => $hash]);
+        $jsonTwo = json_encode(['test_one' => 'one', 'hash' => $newHash]);
+
+
+        // This first snapshot will create the first HTML snapshot.
+        $firstSnapshot = new JsonSnapshot($jsonOne);
+        $firstSnapshot->setDataVisitor($dataVisitor);
+        $firstSnapshot->assert();
+        $snapshotFileName = $firstSnapshot->snapshotFileName();
+        $this->unlinkAfter[] = $snapshotFileName;
+
+        // This second snapshot will compare new data to the existing one.
+        $secondSnapshot = new JsonSnapshot($jsonTwo);
+        $secondSnapshot->setDataVisitor($dataVisitor);
+        $secondSnapshot->setSnapshotFileName($snapshotFileName);
+        $secondSnapshot->assert();
+
+        // Expect the test to fail when the data visitor is not used.
+        $this->expectException(AssertionFailedError::class);
+
+        $failingSnapshot = new JsonSnapshot($jsonTwo);
+        $failingSnapshot->setSnapshotFileName($snapshotFileName);
+        $failingSnapshot->assert();
+    }
+}