+++ tests/unit/tad/Codeception/SnapshotAssertions/HtmlSnapshotTest.php
@@ -3,6 +3,7 @@
 namespace tad\Codeception\SnapshotAssertions;
 
 use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\ExpectationFailedException;
 
 class HtmlSnapshotTest extends BaseTestCase
 {
@@ -14,7 +15,7 @@ class HtmlSnapshotTest extends BaseTestCase
         $class = end($classFrags);
         $methodFrags = explode('::', __METHOD__);
         $method = end($methodFrags);
-        $expectedSnapshotFileName = __DIR__.
+        $expectedSnapshotFileName = __DIR__ .
             sprintf(
                 '/__snapshots__/%s__%s__%d.%s',
                 $class,
@@ -37,7 +38,7 @@ class HtmlSnapshotTest extends BaseTestCase
     {
         $htmlSnapshot = new HtmlSnapshot('foo');
         $snapshot = $htmlSnapshot->snapshotFileName();
-        codecept_debug('Snapshot file: '.$snapshot);
+        codecept_debug('Snapshot file: ' . $snapshot);
         $this->unlinkAfter[] = $snapshot;
 
         $htmlSnapshot->assert();
@@ -55,7 +56,7 @@ class HtmlSnapshotTest extends BaseTestCase
         $htmlSnapshot = new HtmlSnapshot('<p>foo</p>');
         $htmlSnapshot->snapshotPutContents('<h2>bar</h2>');
         $snapshot = $htmlSnapshot->snapshotFileName();
-        codecept_debug('Snapshot file: '.$snapshot);
+        codecept_debug('Snapshot file: ' . $snapshot);
         $this->unlinkAfter[] = $snapshot;
 
         $this->expectException(AssertionFailedError::class);
@@ -73,7 +74,7 @@ class HtmlSnapshotTest extends BaseTestCase
         $htmlSnapshot = new HtmlSnapshot('<ul><li>one</li><li>two</li></ul>');
         $htmlSnapshot->snapshotPutContents('<ul><li>one</li><li>two</li></ul>');
         $snapshot = $htmlSnapshot->snapshotFileName();
-        codecept_debug('Snapshot file: '.$snapshot);
+        codecept_debug('Snapshot file: ' . $snapshot);
         $this->unlinkAfter[] = $snapshot;
 
         $htmlSnapshot->assert();
@@ -95,7 +96,7 @@ HTML;
         $htmlSnapshot = new HtmlSnapshot($current);
         $htmlSnapshot->snapshotPutContents('<ul><li>one</li><li>two</li></ul>');
         $snapshot = $htmlSnapshot->snapshotFileName();
-        codecept_debug('Snapshot file: '.$snapshot);
+        codecept_debug('Snapshot file: ' . $snapshot);
         $this->unlinkAfter[] = $snapshot;
 
         $htmlSnapshot->assert();
@@ -144,5 +145,81 @@ HTML;
         $this->expectException(AssertionFailedError::class);
 
         $htmlSnapshot->assert();
+    }
+
+    /**
+     * It should allow adding a visitor
+     *
+     * @test
+     */
+    public function should_allow_adding_a_visitor()
+    {
+        $removeHashLine = static function ($line) {
+            return strpos($line, 'name="hash"') === false;
+        };
+        $dataVisitor = static function ($expected, $current) use ($removeHashLine) {
+            return [
+                implode("\n", array_filter(explode("\n", $expected), $removeHashLine)),
+                implode("\n", array_filter(explode("\n", $current), $removeHashLine)),
+            ];
+        };
+
+        $hash = md5(microtime());
+        $newHash = md5(microtime());
+
+        $this->assertNotEquals($hash, $newHash);
+
+        $htmlOne = <<<HTML
+<!doctype html>
+<html lang="it">
+<head>
+<meta charset="UTF-8">
+             <title>Document One</title>
+</head>
+<body>
+    <input type="hidden" name="hash" value="$hash">
+    <main>
+        <p>Some text</p>
+    </main>
+</body>
+</html>
+HTML;
+
+        $htmlTwo = <<<HTML
+<!doctype html>
+<html lang="it">
+<head>
+<meta charset="UTF-8">
+             <title>Document One</title>
+</head>
+<body>
+    <input type="hidden" name="hash" value="$newHash">
+    <main>
+        <p>Some text</p>
+    </main>
+</body>
+</html>
+HTML;
+
+
+        // This first snapshot will create the first HTML snapshot.
+        $firstSnapshot = new HtmlSnapshot($htmlOne);
+        $firstSnapshot->setDataVisitor($dataVisitor);
+        $firstSnapshot->assert();
+        $snapshotFileName = $firstSnapshot->snapshotFileName();
+        $this->unlinkAfter[] = $snapshotFileName;
+
+        // This second snapshot will compare new data to the existing one.
+        $secondSnapshot = new HtmlSnapshot($htmlTwo);
+        $secondSnapshot->setDataVisitor($dataVisitor);
+        $secondSnapshot->setSnapshotFileName($snapshotFileName);
+        $secondSnapshot->assert();
+
+        // Expect the test to fail when the data visitor is not used.
+        $this->expectException(AssertionFailedError::class);
+
+        $failingSnapshot = new HtmlSnapshot($htmlTwo);
+        $failingSnapshot->setSnapshotFileName($snapshotFileName);
+        $failingSnapshot->assert();
     }
 }