+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- data visitor support in all assertions to allow pre-processing of the expected and current values before assertions
 
 ## [0.2.2] - 2019-09-21
 ### Added